move defcfg-related code to a new file
